- Removed libc_p from the profile options.  If you don't have this library and attempt to compile with profiling enabled, configure complains that the compiler is "unable to produce executables"!!! To resolve this, we should probably check for libc_p prior to setting it in LDFLAGS.
- Set the optimisation level in profiling to O2.  Previously, we were
profiling un-optimised code.

- Because we weren't setting values for CXXFLAGS or CFLAGS, the
AC_PROG_CXX macro was setting these to default values ("-g -O2") which
appeared with *every* build, regardless of other settings.  We seem to
be using AQSIS_CXXFLAGS and AQSIS_CFLAGS instead of these variables, so
I've set them to null values prior to AC_PROG_CXX so that they don't get
filled in with the defaults.