Fix bug #141074. The transformation of RiPointsPolygons was broken due to the work underway to support autoshadows.
The polygon points class had a check to ensure that it was only transformed once. This was because the individual polygons used to pass their Transform() calls through to the points, and if a transform was applied to a more than one of the member polygons, the points would have been transform twice, incorrectly.

Now, the points can be transformed as many times as required, but the Transform() call on the individual polygons causes an error and does nothing. This is safe because the individual polygons don't enter the pipeline until the first split, at which point the rendering has begun, so transforming the surfaces is invalid anyway.