this is the code from Romain which was a bit modified by me to be like aqsis style guideline.
it is the implementation for blobby and its polygonizer; right now some caveats are there. divide, subtract doesn't work, segments are not working  either. 
remain to implement repression plane and constant blobby.
some work need to be done with the polygonizer itself in particular anything around the bounding box, 
yet another polygonizer could be good too and the equation (implicit_functor) could be exact and give as DSO later.