Revert the camera motion blur handling to the original, more appropriate way.
Camera transformation is stored in a standard CqTransform, along with any transformation motion. At WorldBegin, the current transformation is cleared and the camera transformation stored. When calculating bounding boxes for primitives, the AdjustBoundForTransformationMotion function on the base class now also takes into account any movement imparted by a moving camera. Similarly, when projecting the micropolygons during the Split() function, camera transformation motion is acounted for at the same time as object transformation motion.

This makes more sense, and is essential for the support of automatic rendering passes. Before, the 'main' camera motion was applied to the objects, which meant during rendering of auto passes, like autoshadows, the objects would be moving accordind to the camera motion which doesn't make sense for a shadow pass.