Fix issues with incorrect return values from the ambient occlusion shadeop.
There was some code in there that apparently was supposed to prevent the negative check for cosangle when used in a surface shader. Firstly I cannot understand why this is necessary, commit comments for the change don't explain this. Secondly, the check for if it is a surface shader was odd, checking if a transformation is identity, unreliable and invalid.