Changed the way renderer and shader compiler commands are specified in the Renderer class. Instead of a single string, which may contain arguments, they are now a string list, first string is the command, subsequent strings the optional arguments. This is because the previous method breaks on Unix systems when the shell=True argument is passed to subprocess.Popen.
With the old system, the command line worked out to be...

/bin/sh -c ""aqsis --nocolor" test.rib"

which failed because "aqsis --nocolor" doesn't exist.

Also, modified the way the command is passed to Popen, as on Unix 
systems, if the 'args' property is a list, the first value is the 
command line (in it's entirety), and subsequent values are arguments to 
the shell, not to Aqsis. i.e. 

subprocess.Popen(["aqsis", "test.rib"]...)

results in

/bin/sh -c aqsis test.rib

where test.rib goes to the shell not Aqsis, and thus Aqsis sits there 
waiting for stdin input. The way to fix this is to use 
subprocess.list2cmdline to combine the array of args into one before 
calling Popen.