Fix the rendering of autoshadows.
There was a problem with the order in which the options were stored
and restored, and when the camera transformation was applied.

Now the camera transformation is done outside the RenderWorld call
as part of the function that is requesting the render. This ensures
that the options are pushed before changing the transformation, and
popped before restoring it. The reason for this is that the code that
sets up the camera needs access to the ScreenWindow etc. options.