Refactor the library structure to avoid the use of static libraries, to now use shared libraries for all the shared functionality. This should help us to comply with stringent Fedora conditions.
libaqsis.so - as before, but now without the aqsistypes functionality 
(see below).
libaqsistypes.so - all common functionality, required for various 
things, including the developer libraries.
libslxargs.so - as before, libaqsistypes.so needs to be linked in when 
using.
libri2rib.so - as before, libaqsistypes.so needs to be linked in when 
using.
libshadervm.so - new combined shader vm shared library, including the 
old libshaderexecenv.a, used by libslxargs.so, and libaqsis.so.

Previous other static libraries are removed, and the object files are 
directly linked into the shared libraries/applications that need them, 
thus avoiding the mixing of static and shared compilation that is a 
problem for gcc.