Fix bug #1638918.
The mipmapping code was storing the resulting texturemap buffers in the m_apFlats section, when they should have been stored in the m_apMipMaps section of the caching data. The result is that when a "non-converted" texture was flushed from the cache, the m_apFlats list was cleared (clearing all the mipmaps) but only the first m_apLast entry was cleared. So when a new request for data from that texture came along, the m_apLast entry for the requested MipMap level pointed to a deleted buffer.