Fix for bug #1989897 - Particle Problem
There were various causes of the crash in this bug report, all revolving around the shaderexecenv and the custom grid used for points primitives. The relationship between the number shading points and the number of micropolygons is different for points, but the shaderexecenv was making some assumptions, when calculating derivatives, about the structure of the grid.tex

Applied fixes to the appropriate deriviative calculations to take into accound if the grid comes from a points primitive.