Fix for linux packaging to allow safe removal of the install rpath
We always need a build-time rpath (or LD_LIBRARY_PATH hack) to enable aqsl to
compile the standard shaders at build time.  This means that it's not possible
to turn off the rpath entirely with CMAKE_SKIP_RPATHS.  Nevertheless, this is
desirable for linux packaging.  It's relatively messy to set the
LD_LIBRARY_PATH in a system-independent way, so I opted for doing things via
the build-time rpath.

* Consolidate INSTALL_RPATH stuff into the main cmake file by simply setting
  the global variable CMAKE_INSTALL_RPATH.
* Introduce a new cache variable, AQSIS_USE_RPATH.  Setting this to OFF will
  enable linux packagers to rely on the system linker path rather than
  compiled-in paths.
* Make sure that the CMAKE_SKIP_RPATHS variable isn't settable by the user - it
  will break the build.