Bugfix #1737642: Sides causes odd lighting  (actually due to orientation problems)
This bug was due not to lighting, but to the incorrect orientation of the
surface normals for spheres, cylinders and disks when the current transforation
changed the handedness.  Due to the special way that the normals for quadrics
are computed in *object* space, we shouldn't explicitly flip the normals based
on the handedness of the current transformation.  The only reason to flip the
normals is if the *current orientation* says to do so.

* Make sure the normals are correctly flipped in CqQuadric::DiceAll().

* Remove the CanGenerateNormals() and GenerateGeometricNormals() functions,
  since they are obsoleted by a more general and efficient mechanism inside
  DiceAll().  The only place these were called was for the sphere, cylinder and
  disk.  Since the old mechanism was broken, all three of these gprims suffered
  from incorrectly flipped normals under handedness-changing transformations.