Bugfix #2022216: bicubic patches sometimes cause bad_alloc exception
This was due to integer overflow in the dicable check when rendering with
rather high resolutions.  This caused the patch to be diced at a very high
resolution, leading to an out-of-memory exception.

The solution was to modify the dicable check (a little kludgy, but minimally
invasive for the 1.4 release) and add some extra code to avoid overflow.

Thanks to Paul for pointing out the fix.