Bugfix: Texture wrapping incorrect when texture isn't multiple of tile size
Images which had dimensions which weren't a multiple of image tile size were
incorrectly filtered for texture coordinates outside the texture bounds.  This
was because the pixel iterator CqTileArray::TqIterator didn't correctly check
whether it was empty in the general case where the texture tiles fall off the
image edge.