Bugfix: Ensure that AQSIS_FLTK_INCLUDE_DIR is only used when AQSIS_USE_FLTK is true.
This prevents a cmake makefile generation failure in the case that
AQSIS_USE_FLTK is true, but fltk is not found on the system.

Also reformat two status messages for emphasis and accuracy - display *will* be
built when FLTK isn't found, but the the legacy framebuffer part of it won't.