Refactor the IqTextureSampler interface
The older interface only allowed specification of the sample region using a
quadrilateral.  This change modifies the sampler interface, introducing a new
type of sample region, SqSamplePllgram - a parallelogram.  A parallelograms is
much more closely aligned with the type of data EWA filtering needs, since it
directly provides the Jacobian of the image warp.

The old interface has been retained for its generality.

* Add a new function IqTextureSampler::sample(const SqSamplePllgram&, ...)

* Implement the old sample() method in terms of a call to the new.

* Rewrite the CqEwaFilterFactory constructor to accept the new sample region type.

* Modify some texture shadeops to use the new interface.