Bugfix: aqsistex segfaults when opening 2-channel greyscale TIFFs
A naive reading of the TIFF6 spec. suggests that TIFF files with
PHOTOMETRIC_MINISBLACK will always have a single channel per pixel; the
function CqTiffDirHandle::guessChannels() assumed this was the case.  However,
it makes perfect sense to store an alpha channel along with greyscale data for
a total of two channels and the GIMP certainly does this.

This led to aqsistex underestimating the required buffer size and segfaulting
when loading two-channel images.