Bugfix: Fix a memory leak causing CqAttributes to leak
This was due to a smart-pointer link-cycle which needed to be broken:
CqAttributes -> CqShaderVM (held via IqShader interface) -> CqShaderExecEnv -> CqAttributes

The solution taken was to hold the execution environment using a plain pointer
in CqShaderVM, since the shaderVM doesn't seem to need to manage the lifetime
of the associated shader execenv.  A weak_ptr could have been used, but there
were some performance concerns as it's dereferenced continually, and in any
case it seems like m_pEnv only needs to be valid during a call to
CqShaderVM::Execute() or CqShaderVM::Initialise() so we shouldn't need the
extra safety afforded by weak_ptr.