Add the new RIB parser (untested but compiles).
So far, the parser is pretty limited since it can't deal properly with
parameter lists.  I've got to think about how the inline parameter parsing
and symbol table could be made generic (perhaps it goes in aqsistypes).  This
would avoid the current situation where we've got three disparate versions of
the same code(!) - one in the main renderer, one in the RIB parser and another
in libri2rib.

Other minor changes:
* Make lexer and input buffer noncopyable
* Rename BEGIN_ARRAY to ARRAY_BEGIN so that it's more familiar (compare to Ri*Begin)
* Fix some forgotten inline specifiers in headers