Refactor the RIB parameter list parsing code.
The previous approach was clumsy with respect to type safety: on request, the
rib parser would parse the parameter list, wrapping the parameter values in
boost::any to avoid loosing type information.  The request handler code would
then have to unpack these again to recover the type before doing whatever it
wanted with them.

Instead, the new code introduces a replacement interface, IqRibParamList.  This
has methods for each of the three value types which may be encountered when
constructing the parameter list, allowing type inference to be carried out
*once* inside the parser, after which the user code can do what it wants.

* Removed CqRequestParam, TqRiParamList and related confusing implementation
  details in favour of the simpler and more general IqRibParamList interface.
  Updated parser unit tests for this new approach.

* Move interface stuff which would be needed when declaring a request
  handler into its own file, iribrequest.h