Bugfix: Copy-on-write mechanism for CqAttributes was broken.
When retrieving a CqNamedParameterList in CqAttributes::pAttributeWrite(), the
attribute was held by a newly constructed shared_ptr in addition to the shared
pointer held in the lookup table.  Including the local reference, this means
that the use_count() of the shared_ptr was greater than one.

The copy on write mechanism incorrectly assumed that any use_count greater than
one indicated that the attribute was already held by another CqAttributes
object, and copied it to prevent overrwriting.  This means that *every* call to
a get*AttributeWrite() function caused an attribute to be copied.

Thanks to Tristan Colgate for noticing this bug; it may have significant
performance implications.