Bugfix: ndspy.h sometimes defined fixed integer sizes as the wrong width.
Types such as PkDspyUnsigned32 defined in the ndspy.h header were previously
defined in a very non-robust manner depending on some compiler-specific macros.
This sometimes led to these supposedly fixed-size types getting the wrong
bit-width.  In particular, problems with PkDspyUnsigned32 were reported on
solaris and ppc64.

This fix introduces a new mechanism for the typedefs TqUint32 etc from
aqsis_types.h.  In particular, it removes the use of boost/cstdint.hpp and
replaces it with a mechanism mostly copied from boost (use stdint.h if
possible; otherwise guess using macros from limit.h).  It's unfortunate to
duplicate the boost functionality, but we need to do so for C compatibility,
for use in C headers such as ndspy.h.