Bugfix #2128051: "8-bit character encoding missing"
The RIB parser performed no decoding of escape characters embedded in strings,
so things like \n were passed through unchanged.  This included the escaped
octal character encodings \d \dd and \ddd.

This fix takes the string decoding function from the new RIB parser and
backports it to the old one as an interim measure.

In addition, multiple-line strings were not supported, since the lex regular
expression assumed that strings end at the end of a line.  This is also fixed.