Implement initial Circle of Confusion shading rate optimizations
For a given amount of focal blur, the area covered by the bound of a single
micropolygon increases proportionally to the total image area - this causes
quadratic increase in the render time as a function of the total number of
pixels unless extra measures are taken.

This patch implements GeometricApproximation "focusfactor" which causes the
shading rate of each grid to be increased proportionally to the area of the
circle of confusion.  Also included for testing is GeometricApproximation
"focusfactor1" which increases the area of the circle of confusion
proportionally to the radius of the circle of confusion - one of these should
probably be removed before the next release.  For the moment, both of these
optimizations are turned off by default.


Details:

* Access to the shading rate for dicable testing should now be performed using
  the CqSurface::AdjustedShadingRate() function, which does adjustment for
  "focusfactor" (and for "motionfactor" in the future).

* The standard attribute "System" "ShadingRateSqrt" was removed in favour of
  using "ShadingRate" everywhere instead.

* A bug in CqCurve::GetGridLength() which would cause strange effects for
  curves with shading rates not 1 was corrected.

* The standard attribute "System" "GeometricFocusFactor" was added to store
  focusfactor information.