Refactor aqsis performance timers
For any scene in which large amounts of timing statistics are gathered, the
memory usage of the aqsis performance timers was unnacceptable for production
use - this was because a time was accumulated in memory (represented by a
double) for each timing event.  For millions of events this quickly adds up to
several hundered megabytes.  A good example scene is a menger sponge procedural
at level 4 (soon to be comitted to aqsis svn).

There are still time-based performance penalties incurred from using the
timers, but this patch doesn't really address those problems.

Details:

* Removed storage of individual timing samples in favour of accumulating a
  total time on the fly.  This completely resolves the memory usage issues
  surrounding the old timers implementation.

* Significantly stripped down and refactored the whole timing setup to be
  fairly simple and efficient while performing all the timing tasks which are
  actually used elsewhere in aqsis.  There are still several performance
  optimizations which should be included in future patches, in particular
  avoiding the std::map-based lookup of timer classes.

* Removed all OS-specific timing stuff, in favour of using the boost::timer
  class.  This is in fact based around the standard C clock() function, so is
  unlikely to have the accuracy of the old timer implementation on windows.  It
  should provide cross-platform consistency however and the windows-specific
  implementation can be fairly easily reintroduced into the new framework if
  necessary.