Aqsis timer refactoring, stage 3
CqTimerSet previously used a std::map keyed on the timer names.  While this is
nice and flexible, it's not very fast for large numbers of timing events.
Since the set of timers is known at compile time, it's possible just to work
with a fixed-sized array of timers indexed by and enum, making timer lookup
much faster.

This patch provides an 8% speedup for the level 4 menger sponge test scene at
256x256 when timers are turned on.  The timer overhead still represents 17% of
the total time for the test scene, but 10% of this is taken up in the standard
clock() function.

Details:
* Made CqTimerSet into a template based on a fixed number of timers indexed by
  a "class enum".  This is less flexible but significantly faster.
* Convert all uses of TIMER_START, TIMER_STOP and TIME_SCOPE into
  AQSIS_TIMER_START, AQSIS_TIMER_STOP and AQSIS_TIME_SCOPE, and make them use
  the CqTimerSet interface.