Modify the way that the procedural simulates blender's "end rough"
Instead of making the procedural displace the curve itself, it's possible to
attach a random uniform vector to each curve and do the displacement in a
shader.  This gives the user all the control from the shading language rather
than having to write any C++.  The shader endrough.sl simulates end rough in a
very similar way to the way blender does it internally.