Bugfix #2167498: "uniform user parameters not working for SDS"
This was due to an error in CqSurfaceSubdivisionPatch::StoreDiceAPVar(), in
which the varying index was used for all primvar classes except facevarying and
facevertex.  Adjusting the function to use CqSurfaceSubdivisionPatch::m_FaceIndex
for class_uniform primvars fixes the problem.  A similar problem for
class_constant was also fixed with this patch.