Increase robustness of math function shadeops with nontrivial domains.
The domain of some builtin math functions in the shading language is only part
of the real line, so some inputs result in NaNs, for example sqrt(-1).  If
these are used during shading (to modify P for instance), NaNs may propagate
through the rendering pipeline and eventually cause a segfault.  Some recent
and upcoming refactoring of the renderer core brings this issue into the open,
with resulting segfaults for some of the SLFunction test cases in the RTS.

The specification for the shading language clearly intends that it be pretty
"safe", for instance the RISpec states that array bounds are checked at
runtime.  In keeping with this, I've made sure that the inputs to all math
functions like sqrt() are now checked for validity before calling the function.
Invalid arguments cause a warning and zero is returned.