Refactor CqMicroPolygon::CalculateTotalBound() implementation and usage.
This function was marked as virtual when in fact it's an implementation detail
of CqMicroPolygon.  The versions of CalculateTotalBound() which were
implemented in the child classes CqMicroPolygonMotion and
CqMicroPolygonMotionPoints were removed because they are never called; the
bound calculations for these classes happen inside the associated AppendKey()
functions instead.  Worse, CqMicroPolygonPoints calculates the bound in a
different way, but CalculateTotalBound() was being called for this class,
effectively calling the base class version which is totally irrelevant.

* Remove CqMicroPolygon::m_BoundCode because it's never used.

* Refactor CqMicroPolygon::CalculateTotalBound() for efficiency and simplicity.

* Make it clear that CalculateTotalBound() is an implementation detail by
  avoiding declaring it virtual.

* Make m_IndexCode a TqInt since a 32-bit integer will have more than enough
  bits to use as the index code.