Refactor: Remove DoF expansion from stored micropoly bounding boxes.
Instead of storing the bounding box fully expanded for DoF, this change causes
the expansion to be computed as needed (which is only inside
CqImageBuffer::AddMPG(), so the computation was moved to there.).  The stored
bounding box now fits tightly around the micropolygon, which will help with the
upcoming bounding box optimizations.

Details:
* Introduce CqMicroPolygon::GetBound() which returns the tight bound rather
  than the total bound.  Adjust all child micropolygon classes to compute the
  tight bound appropriately at initialisation.
* Corresponding changes in CqBucketProcessor::RenderMPG_MBOrDof() to use the 
  tight bound rather than the expanded DoF bound.
* Remove the redundent m_Bound from CqMicroPolygonMotionPoints
* Remove redundent bound-culling code from CqBucketProcessor::RenderMicroPoly()
* Introduce min() max() and compMul() functions for CqVector2D to avoid lots of
  operations which are usually written out by hand.