Implement optimization proposal - "Improving bounding boxes for sampling"
The micropolygon sampling code is one of the major contributors to renderer
workload for depth of field and motion-blurred scenes.  This patch
establishes tighter bounding boxes around each micropolygon which are used to
reject potential sample points early, before the more expensive
point-in-polygon test. 

For depth of field, the sample point is displaced along the direction of the
DoF offset by the range of CoC radii for the micropolygon, and tested against
the tight bounding box.  (The range is represented by a min/max CoC multiplier
pair forming a 2D bouding box.)  It's necessary to deal with a range of CoC
radii, since the micropolygon vertices lie at different depths, so the DoF
displacement may cause distortion of the micropolygon.  Along with some caching
of CoC multipliers, this makes the 250mm.rib test case somewhat more than 40%
faster.

For motion blur, the bounding box is interpolated along the motion segment
before the micropolygon vertices are, and used to reject sample points where
possible.  For the test case petunia.rib, this appears to result in something
like a 20% speedup.  This speedup was aided significantly by removing a
redundent call to CacheHitTestValues().

Change summary:

* Introduce fast bounding box-based hit rejection for DoF and MB and DoF+MB 
  in CqMicroPolygon::Sample() and CqMicroPolygonMotion::Sample().
* Introduce CqMicroPolygon::CacheCocMultipliers() to cache any depth of field
  data which doesn't change between different samples of a given micropolygon.
  Also modify CqHitTestCache to hold the additional cached data.
* Remove an additional unnecessary call to pMPG->CacheHitTestValues() for
  motion blur.  This makes the fast bounding box hit rejection scheme much more
  efficient.
* Remove the unused function CqMicroPolygon::CacheHitTestValuesDof()