Refactor: Remove implicit conversions between vector and color types
This refactoring aims to reduce the coupling between the various vector types.
Previously, implicit conversions were allowed between various combinations of
CqVector2D, CqVector3D, CqVector4D and CqColor, which defeats some of the type
safety and can give unexpected results.  Another consequence is that the vector
types must hide at least some of the implementation details because of the
resulting interdependency between the classes.  This makes it very difficult to
implement the vectors as template classes; something which will be useful in
future refactoring.

Details:
* Introduce a set of vectorCast<>() macros for explicit conversion between the
  various vector types.
* Remove implicit CqVector3D -> CqColor conversion
* Remove implicit CqVector3D -> CqVector4D conversion
* Refactor noise_test.cpp as necessary for CqColor <-> CqVector3D conversion changes
* Remove implicit CqColor -> CqVector3D conversion
* Make CqVector2D -> CqVector3D conversion explicit
* Remove implicit CqVector4D -> CqVector3D conversion
* Remove unused operator+ and operator- for "adding" and "subtracting"
  CqVector4D from a CqMatrix.  The same thing can be achieved more explicity in
  other ways.
* Make CqVector3D and CqColor -> CqVector2D conversion explicit.

* All the above changes resulted in a lot of refactoring, necessary to remove
  all implicit conversions and replace them with uses of vectorCast<>().