Futher cleanups to the bucket processing code.
In order to properly support caching of parts of a bucket, 
I've introduced a third region to the bucket processor class.
The original 'S' and 'D' regions have been renamed to make
their purpose clear and a third added, giving...

 * DisplayRegion, the region that is used to define the bucket
   area that is sent to the display.
 * DataRegion, the whole area covered by the samples for this
   bucket.
 * SampleRegion, the area that will be considered for the 
   micropolygon samples.

Removed unused m_aDatas and m_aCoverages member variables on
the bucketprocessor. One wasn't used, and one was entirely
local to the FilterBucket function. 

Finally, added the validity check from #2662 to the static
version of the render micropoly function, this was causing 
a crash in the cropping test.