Bugfix #2644580: "Function overloading in RSL"
The problem was that the scanner places all symbols into a big global symbol
table, relying on unique prefixes to uniqueify variables in different scopes
with the same name.  Unfortunately the scope prefix for a function was just the
function name, and this isn't unique for overloaded functions.  This meant that
a variable "a" in a function "float foo()" could clash with a variable of the
same name "a" in an overloaded function "vector foo()".

The solution taken here was to add a unique global integer ID to every new
scope.  This isn't ideal since the generated code will now can depend on the
order in which functions appear.  It is however an easy way to work around the
problem rather than redesigning the symbol table mechanism completely.