Reimplement pooling for sample hit storage
The original global pooling mechanism was removed during the multithreading
work, but the result is severe slowdowns for scenes with many semitransparent
surfaces (up to a factor of two or so slower has been observed with
depthcomplexity100.rib in the speed test suite).

Sample hit pooling is essential for reasonable cache performance, so this
change reinstates the pooling in an explicit form:  Each CqImagePixel holds a
std::vector<TqFloat> from which sample hit data is allocated.  The improvement
in cache locality is a general optimisation, and also allows the bucket overlap
cache to work with reasonable efficiency.  Because the hit pool is allocated
within the basic reference counted unit (ie, a pixel), the solution will
hopefully play well with multithreading.

Details:
* Make SqImageSample store an index into the associated CqImagePixel instead of
  a pointer to data allocated on the heap.
* Add methods CqImagePixel::sampleHitData() and CqImagePixel::allocateHitData()
  to access and allocate hit data on a CqImagePixel
* Make SqImageSample into a reference type for the sample data rather than a
  value type with respect to the assignment operator and copy constructor.
* Various changes in parts of the code which used the old SqImageSample.data
  member.
* Minor cleanups to documentation and formatting in various places.