Improved DoF sampling patterns and bugfixes to jittering code.
The DoF offsets were previously computed on a regular grid with a single random
offset for each pixel.  This leads to significantly more noise in the focal
blur than necessary, especially at sharp geometric edges since all the offsets
within a pixel are strongly correlated.

Instead, the DoF offsets now use the same multijittered sample distribution
that is used for the screen-space samples.

While updating the DoF jittering, I noticed that the screen-space jittering was
broken.  In particular the shuffling algorithm was buggy so it has been
replaced with rewritten shuffling code which is also used in the new DoF sample
patterns.

During testing the new non-jittered sample distribution was found to be
terribly buggy for > 1x1 samples per pixel, so it has been fixed and updated as
well.