Optimisations for the occlusion tree to reduce cache misses.
These changes modify the way that the occlusion tree updates the depths of the
leaf nodes.  Previously, references to the samples for each leaf node were
stored in the occlusion class.  When updating the tree, we went through each
sample and gathered the depth for that sample to put in the appropriate leaf
node before doing a single-pass update to the internal tree nodes.

The new method stores the leaf node index with the sample point itself.  When the
sample point is updated with a new occluding sample (and only then!), we update
the associated leaf node of the tree to store the maximum depth of all samples
within the leaf.  This means that we only pay (in cache misses and general
execution time) for the sample points that we actually touch.  The update of
the internal tree nodes remains a cache-coherent sweep of the array holding the
tree.

Because leaf node updates are now done per-hit, the occlusion tree is always up
to date at the leaf level and occluded surfaces will correctly be marked as
occluded.  The decision of whether to update the interior nodes is now inside
to the occlusion tree - it knows when the leaves are out of date due to the
per-hit update scheme.