Optimisation: occlusion culling for early rejection in micropolygon sampling.
If a micropolygon bound lies behind the current opaque hit for a sample
we can occlusion cull the micropolygon for that sample point.  This means we
can avoid the more costly point-in-polygon test because we would just throw out
the result anyway.

In scenes where the occlusion tree is less efficient - such as for large motion
blur or depth of field - this optimisation is quite effective at reducing
rending times (petuia.rib: 10% or so on amd64), and also shows modest
improvements for scenes with normal occlusion (capsules.rib: ~2%).