Add CORE_EXPORT flag, remove SVN revision code.
Add a CORE_EXPORT to the build to allow occlusion
code to export a single function needed by the
test harness.

Remove the code that set the revision number to the
SVN repo revision now that we're on Git. Need to find
a suitable replacement to allow easy identification of
the exact build point.

Export treeIndexForPoint for use by test harness.