include <cstring> in various places for g++-4.4
The g++-4.3 and 4.4 releases have cleaned up various header files so that they
no longer #include <string.h> or <cstring> internally.  This means that some
aqsis source files sometimes fail to compile with newer g++ versions.  In this
change I've grepped through the entire source base to identify usages of
string.h functions which I think are likely to be problematic, and have
included <cstring> in the appropriate files.