Fully hide CqShaderVM internals with IqShader interface.
The IqShader abstraction around CqShaderVM was 'leaky' since in order to use
CqShaderVM it was necessary to include the full class definition and associated
implementation details.  This change introduces some factory functions to
manufacture CqShaderVM instances without leaking the implementation details.
This is in preparation for the source tree refactor.

* Introduce factory functions to avoid leaking the CqShaderVM implementation.
* Refactor renderer.cpp to use the new factory functions and avoid having
  access to CqShaderVM internals.