Revise aqsis version.h macro naming
It's useful for external programs to be able to detect which version of aqsis
they're building against.  To this end, version.h was revised for distribution
as an aqsis header, with the version number macros changed to avoid potential
name clashes.  The corresponding cmake variables were also changed to use
clearer names.

Details:

* Rename VERMAJOR, VERMINOR etc to the unambiguous AQSIS_VERSION_MAJOR, ...
* The STRNAME and STRNAME_PRINT macros are removed.  These were used
  inconsistently throughout the source anyway, and don't really seem necessary.
* In the cmake files, the variables MAJOR, MINOR etc have been replaced with
  more descriptive variables VERSION_MAJOR, VERSION_MINOR, etc
  REVISION was replaced by SCM_REVISION and given the value "?" for the time
  being.