Make libslxargs use IqShader rather than CqShaderVM
libslxargs depended on CqShaderVM explicitly which broke the intended
encapsulation via IqShader.  This implies #including "shadervm.h" from slx.cpp
which exposes the CqShaderVM implementation to the world and would require
several private header files to be public.

The remedy taken here is to enhance IqShader slightly to cover the two
additional features needed by libslxargs.  This necessitates modifying
CqLayeredShader as well which is unfortunate, since it can't provide these
functions robustly and shouldn't need to.  Fixing this little niggle would
require significant work however, so for now we do the simplest thing.