Fix a couple of small problems with the Win32 build.
Recent changes have broken the Win32 build in various ways.

1) The removal of the in source unistd.h caused some problems
as the flex provided in win32libs required it. Luckily it also
provided it, so the include paths were updated to point to it.

2) Some symbols were no longer being exported, but were needed by
the testing framework, they have been reinstated while we
investigate a better way of doing the tests.