Fix issue with building shaders under Windows.
The shader compiler needs to find the aqsis_util.dll
and aqsis_slcomp.dll files at build time to be able
to run the sl compiler. This change makes that possible
in the VC IDE.

Not happy about having system specific stuff in the
CMakeLists.txt file of shaders, but couldn't think
of a neat way of doing it quickly, any ideas welcome.