Add shortcut key for deletion of images from books
This change makes CTRL+delete or CTRL+backspace delete an image from the
current book.  It would be better if this was done from within the book browser
handler, but the current code layout makes this very difficult to do while
correctly updating the various widgets which display information about the
image.