Add display of z-buffers to piqsl
Piqsl will now assume that any image with a single float32 channel contains
z-buffer data, and display it in linear greyscale.  The display is scaled such
that the most distant parts of surfaces are black, and closest surfaces white.

Near and far clipping planes are taken into account for an initial estimate of
how the scaling should go, and the scaling is updated at the end after all
buckets have been sent to the framebuffer.

The depth->greyscale mapping is processed on-the-fly, but seems to have
quite acceptable performance.