Turn on the focusfactor approximation for DoF by default.
The scaling factor for in front of the CoC radius was changed from 0.25 to
0.025 (yes, a factor of 10!) based on some experiments.  The aim of this is
that using focusfactor==1 should be almost indistinguishable from the
focusfactor==0 (ie, ideal) result, even in tricky cases.

This does have some performance impact over using larger scaling factors, but
results in much better visual quality in the hard cases.  Of course, users
still have the freedom to adjust the speed/quality tradeoff to their liking.