Add "Matte with Alpha" handling, an extension to RiMatte().
The standard RiMatte function doesn't allow for an easy way to render shadows
cast by CG objects which fall onto live-action parts of a scene.  "Matte with
Alpha" objects adds an extension (usage: RiMatte(RI_MATTEALPHA)) which allows
this to be done more easily.  Matte alpha objects are opaque from the point of
view of the hider, but retain both their colour and opacity in the final image
for maximum flexibility.