Bugfix: Option "limits" "zthreshold" was broken.
This has been broken for a long time - at least since aqsis-1.4 and possibly
much earlier.

There were two distinct problems which caused zthreshold opacity cutoff
computations to fail:
- An "optimisation" which culled micropolygons *before* shading based on their
  Os values was wrong since the shader is free to change the output opacity.
  The offending code was removed.
- The Oi variable was deleted from any grids when fDisplayNeeds() reported that
  it was unneeded.  However, it's usually needed for zthreshold type
  computations, even with the default value of zthreshold.

The meaning of zthreshold was slightly modified for compatibility with 3delight
(and hopefully other renderman renderers) so that all rather than only some
opacity components need to be greater than the threshold for the surface to be
considered opaque.

Associated code in CacheOutputInterpCoeffs*() was also changed to avoid calling
fDisplayNeeds, and instead infer the presence of Oi from the grid itself.