Fix dubious transparency culling based on Os value.
Surface shaders are generally free to set the true surface opacity, Oi, in a
manner completely independent from the base opacity Os, so Oi is the
appropriate opacity variable with which to perform transparency culling of
micropolygons.  Also included in this change are minor rearrangements to bring
some variable definitions closer to where they are used.