Fix z-buffers display so that midpoint filtered buffers show correctly.
Ignore all depths >= FLT_MAX/2 so that if midpoint depth filtering has averaged
the background at depth FLT_MAX with something in the scene (close to 0), that
number won't be considered for the purposes of normalising the depth buffer.
This change allows us to correctly see things in the foreground of the scene.