Bugfix 2811298: Passing varying strings between shaders doesn't work.
The code used to assign values to "varying string" shader variables was
completely broken - it appeared to simply resize the varying array to zero.

Also fix CqShaderVM::GetVariableValue() so that it checks the variable types,
array lengths and varying/uniform conversion are valid before trying to convert
the values held in the associated IqShaderData.  This means the passing of
variables between shaders will be much more pedantic, as specified in the
RISpec.