Implement proper inverse bilinear interpolation for smooth shading interpolation
The old method worked by constructing a linear approximation to colour and
opacity in *screen space*.  The new method works by inverting the (x,y)
coordinates of a hit point into (u,v) coordinates on a micropolygon, and using
these for proper bilinear interpolation of the colours.

The inverse bilinear mapping method chosen at this stage is two iterations of
Newton's method, which seems to be more numerically stable and about as fast as
the analytical solution.  The inverse bilinear mapping from raster (x,y)
coordinates to (u,v) coordinates for a particular micropolygon is encapsulated
by the new functor class CqInvBilinear.

New operations for 2D vectors to support the changes:
  * cross() - cross product function.
  * maxNorm() - the max norm (max of absolute value of components)

In this patch, the reverse bilinear lookup functor is a global static functor,
which gets updated for each micropolygon.  This is a temporary and ugly hack
which will go away when the functor is added to one of the cache structures.