Implement smooth shading for motion blur and depth of field.
This change makes use of the newly added inverse bilinear lookup code to look
up the (u,v) coordinates for each micropolygon hit.  These coordinates are
passed back from the CqMicroPolygon::Sample() function in the same way that the
micropolygon depth is.  The InterpolateOutputs() function then uses the uv
coordinates (rather than xy coordinates in the image plane) to smoothly
interpolate colours across the face of the micropolygon.  This method works
very naturally, regardless of whether the surface undergoes motion or focal
blur.

As part of the changes, the depth of each micropolygon hit is now computed
using the new more accurate bilinear scheme rather than the old method which
used a linear approximation in the image plane.

A bug remains: counter-clockwise and clockwise micropolys get a different order
for vertices as passed into CacheHitTestValues().  This results in incorrect
interpolation for one of the handednesses.