Bugfix to trim curve handling created by smooth interpolation changes.
The ReverseBilinear() function no longer worked after the recent changes, and
is unnecessary in any case because of the new code which computes the (u,v)
coordinates by default.  Given that, ReverseBilinear() was removed, along with
the Point{A,B,C,D} functions and the GetCodedIndex() function.  The
IsDegenerate() function was also removed, given that it was only used in one
location.

A new function, CqMicroPolygon::GetVertices() was added to allow the
micropolygon dumping code to still have access to the vertices.