Optimizations for RiPoints: improve KDTree splitting.
The KDTree class for holding RiPoints was very performing the splitting very
inefficiently.  The main problem was probably that the comparison functor
passed to std::sort was calling several virtual functions for each comparison
in order to look up the point position array.

A secondary concern is that for constructing a KDTree, no actual sorting is
required; we merely need to partition the current set of points about a median.
There exists a linear-time algorithm for doing this in the STL: the
std::nth_element algorithm.  Given this, the KDTree class was modified to make
use of KdTreeData::PartitionElements rather than SortElements, and the
SortElements() method deleted.

Together these optimizations result in roughly a 30% speedup for scenes with
large numbers of particles.  (For example, Matthias' fireworks scene with the
stars (bilinear patches) commented out and sparkles (particles) remaining.)